campaign-results/requirements.txt
Flask>=2.0.1
flask_sqlalchemy>=2.5.1
flask_cors>=3.0.10
python-dotenv>=0.19.0
numpy>=1.21.2
scipy>=1.7.1
pytz>=2021.1
psycopg2-binary>=2.9.1
rollbar>=0.15.2
